pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gwc_datalib"]  

[tool.hatch.build.targets.sdist]
exclude = ["notebooks", "tests", "media"]

[project]
name = "gwc_datalib"
version = "0.0.1"
dependencies = [
    "azure-storage-blob>=12.25.1",
    "numpy",
    "pandas>=2.2.3",
    "python-decouple>=3.8",
    "rasterio>=1.4.3",
    "rioxarray>=0.19.0",
    "xarray>=2025.4.0",
]

authors = [
  { name="Marie Lund Larsen", email="mlla@dhigroup.com" },
]
description = "Useful library"
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Development Status :: 2 - Pre-Alpha",
    "Operating System :: OS Independent",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering",
]

[project.optional-dependencies]
dev = [
  "build",
  "pytest",
  "ruff",
  "mypy",
  "mkdocs",
  "mkdocstrings[python]", 
  "mkdocs-material",
]

test= [
  "pytest",
  "pytest-cov",
  "mypy",
]

notebooks = [
  "jupyter",
]

[project.urls]
"Homepage" = "https://github.com/DHI/gwc-datalib"
"Bug Tracker" = "https://github.com/DHI/gwc-datalib/issues"

[tool.ruff]
# ignore long lines
lint.ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
warn_unreachable = false
no_implicit_optional = true
